Compile time CMUDict with phf_codegen
This is a rather large change the warrants a `2.0` semver change.
We'll change up a small amount of API surface area anyway.

This breaks up the crate into multiple sub-crates, one of which is
responsible for codegen'ing a static, fully-hydrated cmudict.

This will save an enormous amount of server startup time.